requirements.txt
Flask==1.0.2
python-telegram-bot==11.1.0
urllib3==1.24.1